+++ router/index.js
@@ -1,3 +1,11 @@
 const payViaRoutes = require('./pay_via_routes');
+const probeForRoute = require('./probe_for_route');
+const subscribeToPayViaRoutes = require('./subscribe_to_pay_via_routes');
+const subscribeToProbe = require('./subscribe_to_probe');
 
-module.exports = {payViaRoutes};
+module.exports = {
+  payViaRoutes,
+  probeForRoute,
+  subscribeToPayViaRoutes,
+  subscribeToProbe,
+};